front: rename PLS layers files
