+++ front/src/common/Map/Layers/KPLabel.ts
