+++ front/src/modules/simulationResult/components/SimulationResultsMap.tsx
@@ -52,7 +52,7 @@ import {
 import { LAYER_GROUPS_ORDER, LAYERS } from 'config/layerOrder';
 
 import 'common/Map/Map.scss';
-import SNCF_PSL from 'common/Map/Layers/extensions/SNCF/SNCF_PSL';
+import SNCF_PSL from 'common/Map/Layers/extensions/SNCF/PSL';
 import IGN_BD_ORTHO from 'common/Map/Layers/IGN_BD_ORTHO';
 import IGN_SCAN25 from 'common/Map/Layers/IGN_SCAN25';
 import IGN_CADASTRE from 'common/Map/Layers/IGN_CADASTRE';
