+++ front/src/modules/trainschedule/components/ManageTrainSchedule/Map.tsx
@@ -37,7 +37,7 @@ import colors from 'common/Map/Consts/colors';
 import { useMapBlankStyle } from 'common/Map/Layers/blankStyle';
 import { LAYER_GROUPS_ORDER, LAYERS } from 'config/layerOrder';
 import 'common/Map/Map.scss';
-import SNCF_PSL from 'common/Map/Layers/extensions/SNCF/SNCF_PSL';
+import SNCF_PSL from 'common/Map/Layers/extensions/SNCF/PSL';
 import IGN_BD_ORTHO from 'common/Map/Layers/IGN_BD_ORTHO';
 import IGN_SCAN25 from 'common/Map/Layers/IGN_SCAN25';
 import IGN_CADASTRE from 'common/Map/Layers/IGN_CADASTRE';
